First iteration of getting the summary from a wikipedia article. Uses Dagger hilt for dependency injection. Uses retrofit for network related operations. Uses Gson to parse Json data.
The current iteration yields the summary of a "Shoulder" from its Wikipedia Article.
From the Wikimedia Rest Api, the Summary can be accessed from a Json file.
It currently is not possible to directly get the entire Wikipedia article directly from the Wikimedia Rest Api. The closest that can be done is by getting the entire HTML response file, and then parse that response.

It is possible to present the contents of a HTML file with WebView, which either can access a URL directly or by loading in a HTML file from the assets folder. When using the URL directly, it opens up the URL in a Web browser outside of the app.

There is currently one implemented Rest Service which has been implemented, which is the Wikipedia Rest Service.
Currently, the Youtube Rest Api has not been implemented yet.