Started work on a HTML parser class using Jsoup.
The idea is to parse a html document which is downloaded from Wikipedia using Wikimedia:s API.

The standard Html class from Android studio does not work for us, it does not parse complex Html elements and tags properly.

The current idea is to parse an entire HTML document, which is stored in a String. The new MyHtmlParser helper class will be used to parse data. This data will later be presented in a view. The first plan is to present Html paragraphs and sections as Text inside a view, and maybe tables if possible.