New implementation with new screens.
Two new screens: WikipediaScreen and VideoScreen. A navigation class is used to navigate between the screens.

WikipediaScreen presents Wikipedia data.

VideoScreen presents two videos for a specific MyDocument instance with specific VideoIds. When for example "shoulder" is selected as the active document, the drop down menus will present the VideoIds for both regular training and rehab training.

The VideoIds are stored in two dropdown menus as the raw VideoIds, represented as strings. When a specific VideoId is selected, a youtube video is presented through a WebView.