New implementation for WikipediaScreen and VideoScreen.
In the WikipediaScreen,of a Wikipedia articles text and sections are displayed. It  consists of sections (HTML tags h1,h2,h3,h4) and paragraphs (HTML tags p, td and td).

In the VideoScreen, the video titles and the video url thumbnails are displayed in dropdown menus. When a thumbnail is pressed, a WebView is yielded, which can be used to view a Youtube video.